implement authentification

+++ microservice-user_management/src/main/java/org/efrei/UserController.java
@@ -1,5 +1,6 @@
 package org.efrei;
 
+import org.efrei.DAO.UserRepository;
 import org.efrei.entity.UserDetail;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -8,28 +9,45 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequestMapping("/users")
 public class UserController {
+
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private UserRepository userRepsitory;
+
+    private boolean isUserLoggedIn(Long userId) {
+        return userRepsitory.findById(userId).get().getIsConnected();
+    }
+
     @GetMapping("/{userId}")
-    public ResponseEntity<UserDetail> getUserDetail(@PathVariable Long userId) {
+    public ResponseEntity<?> getUserDetail(@PathVariable Long userId) {
+        if (!isUserLoggedIn(userId)) {
+            return ResponseEntity.status(401).body("User not logged in");
+        }
         return ResponseEntity.ok(userService.getUserDetail(userId));
     }
 
     @PutMapping("/{userId}/update")
-    public ResponseEntity<UserDetail> updateUserDetail(@PathVariable Long userId, @RequestBody UserDetail userDetail) {
+    public ResponseEntity<?> updateUserDetail(@PathVariable Long userId, @RequestBody UserDetail userDetail) {
+        if (!isUserLoggedIn(userId)) {
+            return ResponseEntity.status(401).body("User not logged in");
+        }
         return ResponseEntity.ok(userService.updateUserDetail(userId, userDetail));
     }
 
     @DeleteMapping("/{userId}/remove")
-    public ResponseEntity<Void> removeUser(@PathVariable Long userId) {
+    public ResponseEntity<?> removeUser(@PathVariable Long userId) {
+        if (!isUserLoggedIn(userId)) {
+            return ResponseEntity.status(401).body("User not logged in");
+        }
         userService.removeUser(userId);
         return ResponseEntity.ok().build();
     }
 
     @GetMapping("/ping")
     public String ping() {
-        return "User Management service is running correctly !";
+        return "User Management service is running correctly!";
     }
 
     @GetMapping("/pingAuthentification")
